requirements.txt
black==23.12.1
coreapi==2.3.3
Django==5.0.1
djangorestframework==3.14.0